added MLPs implemented in torch

+++ src/sklearn_extensions/preprocessing/lead_time_target.py
@@ -1,3 +1,4 @@
+import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
 
 class LeadFeatureTransformer(BaseEstimator, TransformerMixin):
@@ -15,7 +16,7 @@ class LeadFeatureTransformer(BaseEstimator, TransformerMixin):
 
         self.mask_ = np.arange(ntimesteps) >= self.lead_time
 
-        windowed_features = np.full((ntimesteps, nfeatures, self.lead_time), np.nan)
+        windowed_features = np.full((ntimesteps, nfeatures, self.lead_time), np.nan, dtype=float)
         for idx, val in enumerate(X[:-self.lead_time]):
             windowed_features[idx + self.lead_time] = X[idx : idx + self.lead_time, :]
         
@@ -30,7 +31,7 @@ class LeadFeatureTransformer(BaseEstimator, TransformerMixin):
         if self.apply_mask:
             y_result = y[self.lead_time:]
         else:
-            y_result = np.full_like(y, np.nan)
+            y_result = np.full_like(y, np.nan, dtype=float)
             y_result[self.lead_time:] = y[:-self.lead_time]
 
         return y_result
